+++ taskList.txt
@@ -1,6 +1,5 @@
 Here are the tasks in your list:
 1.[T][X] read book
 2.[E][X] soccer match (at: Monday 3pm)
-3.[D][X] homework (by: Tuesday 5pm)
-4.[D][ ] math homework (by: Mar 02 2019 22:00)
-5.[E][ ] charity match (at: Mar 02 2024 23:15)
+3.[D][ ] math homework (by: Mar 02 2019 22:00)
+4.[E][ ] charity match (at: Mar 02 2024 23:15)
